simple_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
python-dotenv==1.0.0
numpy==1.24.3
pydantic==2.5.0